+++ plugins/button/lang/sq.js
@@ -1,8 +1,0 @@
-/**
- * @license Copyright (c) 2003-2018, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-
-CKEDITOR.plugins.setLang( 'button', 'sq', {
-	selectedLabel: '%1 (Përzgjedhur)'
-} );