+++ plugins/clipboard/lang/af.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'af', {
 	paste: 'Byvoeg',
 	pasteNotification: 'Druk %1 om by te voeg. You leser ondersteun nie die toolbar knoppie of inoud kieslysie opsie nie. ',
 	pasteArea: 'Area byvoeg',
-	pasteMsg: 'Voeg jou inhoud in die gebied onder by en druk OK',
-	title: 'Byvoeg'
+	pasteMsg: 'Voeg jou inhoud in die gebied onder by en druk OK'
 } );