+++ plugins/clipboard/lang/el.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'el', {
 	paste: 'Επικόλληση',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Περιοχή Επικόλλησης',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Επικόλληση'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );