+++ plugins/clipboard/lang/fa.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'fa', {
 	paste: 'چسباندن',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'محل چسباندن',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'چسباندن'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );