+++ plugins/clipboard/lang/fr-ca.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'fr-ca', {
 	paste: 'Coller',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Coller la zone',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Coller'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );