+++ plugins/clipboard/lang/km.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'km', {
 	paste: 'បិទ​ភ្ជាប់',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'តំបន់​បិទ​ភ្ជាប់',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'បិទ​ភ្ជាប់'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );