+++ plugins/clipboard/lang/ko.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'ko', {
 	paste: '붙여넣기',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: '붙여넣기 범위',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: '붙여넣기'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );