+++ plugins/clipboard/lang/mk.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'mk', {
 	paste: 'Залепи (Paste)',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Простор за залепување',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Залепи (Paste)'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );