+++ plugins/clipboard/lang/tt.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'tt', {
 	paste: 'Өстәү',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Өстәү мәйданы',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Өстәү'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );