+++ tests/plugins/link/link.js
@@ -592,8 +592,37 @@
 			correctInput: '123456789',
 			incorrectLinkAssertionCallback: assertIncorrectLinks,
 			correctLinkAssertionCallback: assertCorrectLinks
-		} )
+		} ),
+
+		// (#2478)
+		'test Ctrl+K keystroke': assertKeystroke( 75 ),
+
+		'test Ctrl+L keystroke': assertKeystroke( 76 )
 	} );
+
+	function assertKeystroke( key ) {
+		return function() {
+			var bot = this.editorBots.noValidation,
+				editor = bot.editor;
+
+			bot.setData( '', function() {
+				editor.once( 'dialogShow', function( evt ) {
+					resume( function() {
+						var dialog = evt.data;
+
+						dialog.hide();
+						assert.areSame( 'link', dialog._.name );
+					} );
+				} );
+
+				editor.editable().fire( 'keydown', new CKEDITOR.dom.event( {
+					keyCode: key,
+					ctrlKey: true
+				} ) );
+				wait();
+			} );
+		};
+	}
 
 	function assertPhoneLinks( config ) {
 		return function() {